+++ custom_components/home_maintenance/const.py
@@ -4,7 +4,7 @@ import voluptuous as vol
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers import config_validation as cv
 
-VERSION = "1.5.10"
+VERSION = "1.5.11"
 NAME = "Home Maintenance"
 MANUFACTURER = "@TJPoorman"